backend/requirements.txt
Flask==3.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
google-cloud-storage==2.13.0
Flask-Cors==4.0.0
gunicorn==21.2.0
